realizohet lidhja ne mes disa crudave

+++ Application/KohezgjatjaOres/Create.cs
@@ -12,7 +12,7 @@ namespace Application.KohezgjatjaOres
         public class Command : IRequest
         {
         public Guid Id {get; set;}
-        public int kohaMin {get; set;}
+        public float kohaMin {get; set;}
         public float oraNisjes {get; set;}
         }
